1) Improved rockets and sparks for bullet physics    Parameters for rockets etc. can for now be    set in data/*.projectile 2) Fixed bugs causing rockets not to explode (and    potentialy start rotating) 3) Added (somewhat temporarily) some cheats to get    rockets, homing missiles and sparks as F1/2/3 4) Fixed several memory leaks. 5) Fixed (very rare) occurring crash when deleting    phsyics. 6) Started to remove non-bullet physics code 7) Bullet physics is now the default, and it    can not be disabled anymore!! Important note: From this revision on, non-bullet physics is NOT supported anymore, and will (most likely) not compile. For now a -DBULLET is still necessary (it is added as default), but the code will be cleaned up in the near future.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1325 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/kart_properties_manager.hpp
@@ -42,7 +42,7 @@ public:
     const int                getKartId              (const std::string IDENT);
     void                     loadKartData           (bool dont_load_models=false);
     const float              getMaximumSteeringAngle() {return m_max_steer_angle;}
-    const unsigned int       getNumberOfKarts       () {return m_karts_properties.size();}
+    const unsigned int       getNumberOfKarts       () {return (unsigned int)m_karts_properties.size();}
 
     /** Return len random karts */
     std::vector<std::string> getRandomKarts         (int len);
